Promote time-varying layout to full example
