--- examples/python/graph_lattice/README.md
@@ -1,29 +1,0 @@
-<!--[metadata]
-title = "Graph lattice"
-tags = ["Graph", "Layout"]
-thumbnail = "https://static.rerun.io/graph_lattice/f9169da9c3f35b7260c9d74cd5be5fe710aec6a8/480w.png"
-thumbnail_dimensions = [480, 269]
-channel = "main"
--->
-
-This example shows different attributes that you can associate with nodes in a graph.
-Since no explicit positions are passed for the nodes, Rerun will layout the graph automatically.
-
-<picture>
-  <img src="https://static.rerun.io/graph_lattice/f9169da9c3f35b7260c9d74cd5be5fe710aec6a8/full.png" alt="">
-  <source media="(max-width: 480px)" srcset="https://static.rerun.io/graph_lattice/f9169da9c3f35b7260c9d74cd5be5fe710aec6a8/480w.png">
-  <source media="(max-width: 768px)" srcset="https://static.rerun.io/graph_lattice/f9169da9c3f35b7260c9d74cd5be5fe710aec6a8/768w.png">
-  <source media="(max-width: 1024px)" srcset="https://static.rerun.io/graph_lattice/f9169da9c3f35b7260c9d74cd5be5fe710aec6a8/1024w.png">
-  <source media="(max-width: 1200px)" srcset="https://static.rerun.io/graph_lattice/f9169da9c3f35b7260c9d74cd5be5fe710aec6a8/1200w.png">
-</picture>
-
-## Used Rerun types
-[`GraphNodes`](https://www.rerun.io/docs/reference/types/archetypes/graph_nodes?speculative-link),
-[`GraphEdges`](https://www.rerun.io/docs/reference/types/archetypes/graph_edges?speculative-link)
-
-## Run the code
-
-```bash
-pip install -e examples/python/graph_lattice
-python -m graph_lattice
-```
